[feat] GTFS import pipeline: streaming zip reader -> canonical (stops->terminals+stops_canonical, services+calendar/exceptions with id-offset, routes/trips/frequencies by NK, stop_times via pgx.Batch 2K), ServiceID int64 vertical, walk-transfers generation (400m, city feeds have no transfers.txt - network was fragmented), connections sorted at LoadNetwork, per-day network cache in App; live SPb feed: 8715 stops/545 routes/128K trips/3.5M stop_times in 3m39s, SPb routes now routable (Moscovsky vokzal->Nevsky 13:02->13:14)

+++ internal/adapters/mintrans/contract.go
@@ -44,7 +44,7 @@ func ValidateDatasetContract(raw []byte) error {
 	if err := checkStops(ds.Stops); err != nil {
 		return err
 	}
-	services := map[int]bool{}
+	services := map[int64]bool{}
 	for i, svc := range ds.Services {
 		if svc.ID <= 0 {
 			return fmt.Errorf("reestr: services[%d].id = %d, want > 0", i, svc.ID)
@@ -167,7 +167,7 @@ func blockDaysOf(b *reestrBlock) string {
 	return b.Days
 }
 
-func checkSchedules(scheds []reestrSched, routeSet, stopSet map[string]bool, services map[int]bool) error {
+func checkSchedules(scheds []reestrSched, routeSet, stopSet map[string]bool, services map[int64]bool) error {
 	seenDir := map[string]bool{}
 	for i, sc := range scheds {
 		if sc.Route == "" {
